* Improvements to Anaglyph stereo with imaging pipeline:
  Implemented mono-overlay support and a couple of interesting
  new anaglyph algorithms - as well as an additional shader in
  SetAnaglyphStereoParameters.m

  -> Nicely demonstrated with ImagingStereoMoviePlayer.m

* ImagingStereoDemo now supports setup for stereomode 10 on
  OS/X as well (dual display stereo).

* Small fix to AlphaImageDemo to account for small roundoff
  errors in old gfx-hardware which may make the rectangular
  support of the Gaussian mask visible.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@901 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ Psychtoolbox/PsychDemos/AlphaImageDemo/AlphaImageDemo.m
@@ -132,7 +132,7 @@ try
     fprintf('Size  mask texture: %d x %d\n', RectWidth(mRect), RectHeight(mRect));
 
     % Blank sceen
-    Screen('FillRect',w, gray);
+    Screen('FillRect',w, uint8(gray));
     Screen('Flip', w);
 
     % initialize mouse position at center
